create client server, removing imgs from localstorage, adjusting functions, checking before send data to backend
