+++ static/stylesheets/fonts/NovaFlat-Regular.ttf
